requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.9.2
PyBrain==0.3
python-midi==0.2.4
scipy==0.16.0
Werkzeug==0.10.4
wheel==0.24.0
